backend/hunter_agent/requirements.txt
openai
requests
numpy
scikit-learn
python-dotenv 